renaming private module utils to drop vmware prefix
