+++ changelogs/fragments/152-rename-private-module-utils.yml
@@ -1,0 +1,3 @@
+---
+minor_changes:
+  - rename private module_utils to drop the redundant vmware prefix
