+++ plugins/module_utils/_vsphere_tasks.py
