+++ plugins/modules/cluster_info.py
@@ -163,7 +163,7 @@ from ansible_collections.vmware.vmware.plugins.module_utils.argument_spec import
     rest_compatible_argument_spec
 )
 from ansible_collections.vmware.vmware.plugins.module_utils._module_rest_base import ModuleRestBase
-from ansible_collections.vmware.vmware.plugins.module_utils._vmware_facts import (
+from ansible_collections.vmware.vmware.plugins.module_utils._facts import (
     ClusterFacts,
     vmware_obj_to_json
 )
