+++ plugins/modules/vm_portgroup_info.py
@@ -67,7 +67,7 @@ vm_portgroup_info:
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.vmware.vmware.plugins.module_utils._module_pyvmomi_base import ModulePyvmomiBase
-from ansible_collections.vmware.vmware.plugins.module_utils import _vmware_network as vmware_network
+from ansible_collections.vmware.vmware.plugins.module_utils import _network as vmware_network
 from ansible_collections.vmware.vmware.plugins.module_utils._module_rest_base import ModuleRestBase
 from ansible_collections.vmware.vmware.plugins.module_utils.argument_spec import rest_compatible_argument_spec
