+++ plugins/modules/vm_powerstate.py
@@ -229,7 +229,7 @@ from ansible_collections.vmware.vmware.plugins.module_utils._module_pyvmomi_base
 from ansible_collections.vmware.vmware.plugins.module_utils.argument_spec import (
     base_argument_spec
 )
-from ansible_collections.vmware.vmware.plugins.module_utils._vmware_tasks import TaskError, RunningTaskMonitor, VmQuestionHandler
+from ansible_collections.vmware.vmware.plugins.module_utils._vsphere_tasks import TaskError, RunningTaskMonitor, VmQuestionHandler
 
 
 class VmPowerstateModule(ModulePyvmomiBase):
